// frontend/src/api/model/getCommentListAdminParams.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */

export type GetCommentListAdminParams = {
  /**
   * 关键词 (评论内容)
   */
  keyword?: string;
  /**
   * 页码
   * @minimum 1
   */
  page?: number;
  /**
   * 每页数量
   * @minimum 1
   * @maximum 100
   */
  size?: number;
};